+++ src/tests/unit/passwordReset/ResetPassword.test.js
@@ -1,42 +1,42 @@
-/* eslint-disable no-undef */
+import React from 'react';
+import { BrowserRouter as Router } from 'react-router-dom';
 import expect from 'expect';
-import RequestPasswordReset from '../../../store/reducers/requestPasswordReset';
-import { requestResetPasswordSuccess } from '../../../store/actions/requestResetPassword';
+import { mount } from '../../enzyme';
 
-const INITIAL_STATE = {
-  error: '',
-  message: {},
+import { ResetPassword } from '../../../containers/PasswordReset/ResetPassword';
 
+const setUp = () => {
+  const props = {
+    // eslint-disable-next-line no-undef
+    handleSubmit: jest.fn(),
+    // eslint-disable-next-line no-undef
+    resetPassword: jest.fn(),
+  };
+  const wrapper = mount(
+    <Router>
+      <ResetPassword {...props} />
+    </Router>,
+  );
+  return {
+    props,
+    wrapper,
+  };
 };
 
 // eslint-disable-next-line no-undef
-describe('Test RESETREDUCER', () => {
-  it('should return the initial state', () => {
-    const state = RequestPasswordReset(INITIAL_STATE, {});
-    expect(state).toEqual(INITIAL_STATE);
+describe('Request password reset page rendered ', () => {
+  const { wrapper } = setUp();
+  // eslint-disable-next-line no-undef
+  it(' render reset password r form ', () => {
+    const ResetForm = wrapper.find('.form-wrapper');
+    expect(ResetForm.exists())
+      .toBe(true);
   });
 
-  it('should handle get reset password link success', () => {
-    const payload = 'success! reset password link has been sent to your email';
-    const resetSuccessAction = requestResetPasswordSuccess(payload);
-    const expectedState = {
-      ...INITIAL_STATE,
-      message: resetSuccessAction.message,
-
-    };
-    const state = RequestPasswordReset(INITIAL_STATE, resetSuccessAction);
-    expect(state).toEqual(expectedState);
-  });
-
-  it('should handle get reset password link error', () => {
-    const payload = 'Failed email is not registered';
-
-    const resetFailureAction = requestResetPasswordSuccess(payload);
-    const expectedState = {
-      ...INITIAL_STATE,
-      message: resetFailureAction.message,
-    };
-    const state = RequestPasswordReset(INITIAL_STATE, resetFailureAction);
-    expect(state).toEqual(expectedState);
+  // eslint-disable-next-line no-undef
+  it(' test handle onsubmit event', () => {
+    wrapper.find('form').simulate('submit', {
+      preventDefault: () => {},
+    });
   });
 });
